+++ orchestrator_service/src/main/java/com/override/orchestrator_service/service/calc/TransactionHandler.java
@@ -1,7 +1,7 @@
 package com.override.orchestrator_service.service.calc;
 
 public interface TransactionHandler {
-    float calculateAmount(String transaction);
+    double calculateAmount(String transaction);
 
     String getTransactionComment(String transaction);
